+++ src/main.js
@@ -24,7 +24,7 @@ const BASE_URL = 'https://18.ecmascript.pages.academy/big-trip';
 const POINTS_URL = `${BASE_URL}/points`;
 const DESTINATIONS_URL = `${BASE_URL}/destinations`;
 const OFFERS_URL = `${BASE_URL}/offers`;
-const AUTH = 'Basic bo1080bdzbgg';
+const AUTH = 'Basic bo1480bdzbgg';
 
 
 /** @type {Store<Point>} */
