+++ src/view/js/sort-view.js
@@ -1,9 +1,76 @@
-import RadioGroupView, {html} from './radio-group-view.js';
+import View, { html } from './view.js';
 
-export default class SortView extends RadioGroupView {
+export default class SortView extends View {
+  get inputSelector() {
+    return '[type="radio"]';
+  }
+
+  /**
+   * @type {NodeListOf<HTMLInputElement>}
+   */
+  get inputViews() {
+    return this.querySelectorAll(this.inputSelector);
+  }
+
+  getValue() {
+    /** @type {HTMLInputElement} */
+    const inputCheckedView = this.querySelector(`${this.inputSelector}:checked`);
+
+    if (inputCheckedView) {
+      return inputCheckedView.value;
+    }
+
+    return '';
+  }
+
+  /**
+   * @param {string} value
+   */
+  setValue(value) {
+    /** @type {HTMLInputElement} */
+    const inputView = this.querySelector(`${this.inputSelector}[value="${value}"]`);
+
+    if (inputView) {
+      inputView.checked = true;
+    }
+
+    return this;
+  }
+
+  getIndex() {
+    return [...this.inputViews].findIndex((view) => view.checked);
+  }
+
+  /**
+   * @param {number} index
+   */
+  setIndex(index, notify = true) {
+    const views = this.inputViews;
+    const rangeIndex = (views.length + index) % views.length;
+
+    views.item(rangeIndex).checked = true;
+
+    if (notify) {
+      views.item(rangeIndex).dispatchEvent(new Event('change', { bubbles: true }));
+    }
+
+    return this;
+  }
+
+  /**
+   * @param {boolean[]} flags
+   */
+  setOptionsDisabled(flags) {
+    this.inputViews.forEach((view, index) => {
+      view.disabled = flags[index];
+    });
+
+    return this;
+  }
   /**
    * @override
    */
+
   createTemplate() {
     return html`
       <form class="trip-events__trip-sort  trip-sort" action="#" method="get"></form>
